+++ src/pybear/feature_extraction/text/_TextLookup/_shared/_validation/_skip_always.py
@@ -10,23 +10,29 @@ from typing import (
     Sequence,
 )
 
-from ......base._check_1D_str_sequence import check_1D_str_sequence
+import re
 
 
 
 def _val_skip_always(
-    _skip_always: Sequence[str] | None
+    _SKIP_ALWAYS: None | Sequence[str | re.Pattern[str]]
 ) -> None:
-    """Validate skip_always.
+    """Validate `SKIP_ALWAYS`.
 
-    Must be a 1D sequence of strings or None.
+    Must be None or a 1D sequence of strings and/or re.compile objects.
+    Cannot be empty, cannot have duplicate entries.
 
     Parameters
     ----------
-    _skip_always : Sequence[str] | None
-        A 1D sequence of strings that when there is a case-sensitive
-        match against a word in the text, that word is skipped without
-        further action and left in the body of text.
+    _SKIP_ALWAYS : None | Sequence[str | re.Pattern[str]]
+        A non-empty 1D sequence of strings and/or re.compile objects that
+        when there is a match against a word in the text, that word is
+        removed from the body of text.
+
+    Raises
+    ------
+    ValueError
+    TypeError
 
     Returns
     -------
@@ -35,13 +41,37 @@ def _val_skip_always(
     """
 
 
-    if _skip_always is None:
+    if _SKIP_ALWAYS is None:
         return
 
 
-    check_1D_str_sequence(_skip_always, require_all_finite=True)
+    # check_1D_str_sequence(_SKIP_ALWAYS, require_all_finite=True)
 
+    _err_msg = (f"'SKIP_ALWAYS' must be None or a non-empty 1D list-like "
+                f"of strings or re.compile objects. ")
 
+    _addon = ""
+
+    try:
+        list(iter(_SKIP_ALWAYS))
+        if isinstance(_SKIP_ALWAYS, (str, dict)):
+            raise Exception
+        if len(_SKIP_ALWAYS) == 0:
+            _addon = "Got empty."
+            raise UnicodeError
+        if not all(map(
+            isinstance,
+            _SKIP_ALWAYS,
+            ((str, re.Pattern) for i in _SKIP_ALWAYS)
+        )):
+            raise Exception
+        if len(_SKIP_ALWAYS) != len(set(_SKIP_ALWAYS)):
+            _addon = "Got duplicate entries."
+            raise UnicodeError
+    except UnicodeError:
+        raise ValueError(_err_msg + _addon)
+    except Exception as e:
+        raise TypeError(_err_msg + _addon)