+++ src/pybear/feature_extraction/text/_TextLookup/_shared/_validation/_split_always.py
@@ -10,26 +10,34 @@ from typing import (
     Sequence,
 )
 
+import re
+
 from ......base._check_1D_str_sequence import check_1D_str_sequence
 
 
 
 def _val_split_always(
-    _split_always: dict[str, Sequence[str]] | None
+    _SPLIT_ALWAYS: None | dict[str | re.Pattern[str], Sequence[str]]
 ) -> None:
-    """Validate split_always.
+    """Validate `SPLIT_ALWAYS`.
 
-    Must be None or a dictionary with strings as keys and sequences of
-    strings as values.
+    Must be None or a non-empty dictionary with strings and/or re.compile
+    objects as keys and non-empty sequences of strings as values.
 
     Parameters
     ----------
-    _split_always : dict[str, Sequence[str]] | None
-        None or a dictionary with strings as keys and sequences of
-        strings as values. When a key in the dictionary is a
-        case-sensitive match against a word in the text, the matching
-        word is removed and the corresponding words in the sequence are
-        inserted, starting in the position of the original word.
+    _SPLIT_ALWAYS : None | dict[str | re.Pattern[str], Sequence[str]]
+        None or a non-empty dictionary with strings and/or re.compile
+        objects as keys and sequences of strings as values. When a key
+        in the dictionary is a match against a word in the text, the
+        matching word is removed and the corresponding words in the
+        sequence are inserted, starting in the position of the original
+        word.
+
+    Raises
+    ------
+    ValueError
+    TypeError
 
     Returns
     -------
@@ -38,26 +46,40 @@ def _val_split_always(
     """
 
 
-    if _split_always is None:
+    if _SPLIT_ALWAYS is None:
         return
 
+    _err_msg = (
+        f"'SPLIT_ALWAYS' must be None or a non-empty dictionary with "
+        f"strings or re.compile objects as keys and non-empty sequences "
+        f"of strings as values. "
+    )
+
+    _addon = ""
 
     try:
-        if not isinstance(_split_always, dict):
+        if not isinstance(_SPLIT_ALWAYS, dict):
             raise Exception
-        for k, v in _split_always.items():
-            if not isinstance(k, str):
+        if len(_SPLIT_ALWAYS) == 0:
+            _addon = f"Got empty dictionary."
+            raise UnicodeError
+        for k, v in _SPLIT_ALWAYS.items():
+            if not isinstance(k, (str, re.Pattern)):
+                _addon = f"Got bad key."
                 raise Exception
             try:
-                iter(v)
+                list(iter(v))
                 check_1D_str_sequence(v, require_all_finite=True)
             except:
+                _addon = f"\nGot bad sequence of strings for key '{k}'."
                 raise Exception
-    except:
-        raise TypeError(
-            f"'split_always' must be None or a dictionary with strings "
-            f"as keys and sequences of strings as values."
-        )
+            if len(v) == 0:
+                _addon = f"\nGot empty sequence for key '{k}'."
+                raise UnicodeError
+    except UnicodeError:
+        raise ValueError(_err_msg + _addon)
+    except Exception as e:
+        raise TypeError(_err_msg + _addon)