+++ tests/feature_extraction/text/TextLookup/TLRT_10sec_test.py
@@ -9,9 +9,10 @@
 import pytest
 from unittest.mock import patch
 
+from copy import deepcopy
 import io
 import numbers
-from copy import deepcopy
+import re
 
 import numpy as np
 import pandas as pd
@@ -138,9 +139,13 @@ class TestTextLookupRealTime:
             ["CRUMBLE", "WAX"]
     ]
 
+    # END fixtures ** * ** * ** * ** * ** * ** * ** * ** * ** * ** * **
 
 
-    def test_accuracy(self, _kwargs, _X, _exp):
+    def test_accuracy_insitu(self, _kwargs, _X, _exp):
+
+        # all decisions about words not in Lexicon are handled manually
+        # in-situ, nothing is automatic via ALWAYS holders passed at init
 
         TestCls = TLRT(**_kwargs)
 
@@ -210,6 +215,95 @@ class TestTextLookupRealTime:
             ['ZONKING']
         )
 
+
+    def test_accuracy_ALWAYS_init(self, _kwargs, _X, _exp):
+
+        # all decisions about words not in Lexicon are automatic via ALWAYS
+        # holders passed at init, nothing is handled manually in-situ
+
+        _new_kwargs = deepcopy(_kwargs)
+        _new_kwargs['SKIP_ALWAYS'] = [re.compile('ZONK.+')]
+        _new_kwargs['DELETE_ALWAYS'] = [
+            'TORTAGLOOM', 'SNORLUX', 'GLENSHWINK', 'JUMBLYWUMP',
+            re.compile('^QUACK.+$', re.I)
+        ]
+        _new_kwargs['REPLACE_ALWAYS'] = {
+            re.compile('^GLOURY$'): 'GLORY',
+            'BEAUTIFULL': 'BEAUTIFUL'
+        }
+        _new_kwargs['SPLIT_ALWAYS'] = {
+            re.compile('BLOOMTRIX'): ['BLOOM', 'TRIX']
+        }
+
+        TestCls = TLRT(**_new_kwargs)
+
+        # need to do "delete once" and "replace once" on some words so
+        # that they are handled and satisfy stdin, but they dont go into
+        # any of the ALWAYS buckets and the assertions below are satisified.
+        # because we are passing *TRIX* as a new word in SPLIT_ALWAYS
+        # and it is not in Lexicon, need to do a stdin to deal with that.
+        # Add *TRIX* to Lexicon, which will put it in TestCls.KNOWN_WORDS
+        user_inputs = f"d\nd\nd\na\ne\nAMAZING\ny\n\ne\nMAGNIFICENT\ny\nd\nd\nc\n"
+        with patch('sys.stdin', io.StringIO(user_inputs)):
+            out = TestCls.transform(_X)
+
+        for r_idx in range(len(_exp)):
+            assert np.array_equal(out[r_idx], _exp[r_idx])
+
+        nr_ = TestCls.n_rows_
+        assert isinstance(nr_, numbers.Integral)
+        assert nr_ == len(_X)
+        del nr_
+
+        rs_ = TestCls.row_support_
+        assert isinstance(rs_, np.ndarray)
+        assert all(map(isinstance, rs_, (np.bool_ for _ in rs_)))
+        assert len(rs_) == len(_X)
+        assert np.array_equal(rs_, [True] * len(_X))
+        del rs_
+
+        assert np.array_equal(
+            TestCls.LEXICON_ADDENDUM_,
+            ['TRIX']
+        )
+
+        assert TestCls.KNOWN_WORDS_[0] == 'TRIX'
+
+        # this proves that the Lexicon singleton class attribute's
+        # lexicon_ attribute is not mutated when no deepcopy and adding
+        # words to KNOWN_WORDS_ (which is just a shallow copy of lexicon_)
+        assert 'TRIX' not in TestCls.get_lexicon()
+
+        assert np.array_equal(
+            list(TestCls.SPLIT_ALWAYS_.keys()),
+            [re.compile('BLOOMTRIX')]
+        )
+
+        assert np.array_equal(
+            list(TestCls.SPLIT_ALWAYS_.values()),
+            [['BLOOM', 'TRIX']]
+        )
+
+        assert np.array_equal(
+            TestCls.DELETE_ALWAYS_,
+            ['TORTAGLOOM', 'SNORLUX', 'GLENSHWINK',
+            'JUMBLYWUMP', re.compile('^QUACK.+$', re.I)]
+        )
+
+        assert np.array_equal(
+            list(TestCls.REPLACE_ALWAYS_.keys()),
+            [re.compile('^GLOURY$'), 'BEAUTIFULL']
+        )
+
+        assert np.array_equal(
+            list(TestCls.REPLACE_ALWAYS_.values()),
+            ['GLORY', 'BEAUTIFUL']
+        )
+
+        assert np.array_equal(
+            TestCls.SKIP_ALWAYS_,
+            [re.compile('ZONK.+')]
+        )
 
 
     def test_array_all_str_numbers(self, _kwargs):